+++ TRXpense.App.Web/Controllers/CountryAllowanceController.cs
@@ -10,25 +10,23 @@ namespace TRXpense.App.Web.Controllers
     public class CountryAllowanceController : Controller
     {
         private readonly ICountryAllowanceRepository _countryAllowanceRepository;
+        private readonly ITravelReportRepository _travelReportRepository;
 
         public CountryAllowanceController()
         {
             _countryAllowanceRepository = new CountryAllowanceRepository();
+            _travelReportRepository = new TravelReportRepository();
         }
 
         // GET: CountryAllowance
         public ActionResult Index(int? page, string query = null)
         {
-            var allowances = _countryAllowanceRepository
-                .GetAllFromDatabaseEnumerable()
-                .ToList()
-                .MapToViews()
-                .OrderBy(o => o.Country);
+            var allowances = _countryAllowanceRepository.GetAllFromDatabaseEnumerable().ToList().MapToViews().OrderBy(o => o.Country);
 
             // paging
-            int pageSize = 6;
+            int pageSize = 5;
             var pageNumber = page ?? 1; // if no page was specified in the querystring, default to the first page (1)
-            var onePageOfAllowances = allowances.ToPagedList(pageNumber, pageSize); // will only contain 6 products max because of the pageSize
+            var onePageOfAllowances = allowances.ToPagedList(pageNumber, pageSize); // will only contain 5 items max because of the pageSize
 
             // searching
             if (!string.IsNullOrEmpty(query))
@@ -37,6 +35,7 @@ namespace TRXpense.App.Web.Controllers
                 var allowanceSearched = _countryAllowanceRepository.GetAllFromDatabaseEnumerable()
                     .Where(a => a.Country.ToLower().Contains(query.ToLower())
                         || a.AllowanceCurrency.ToLower().Contains(query.ToLower())
+                        || a.OfficialCurrency.ToLower().Contains(query.ToLower())
                         || a.Amount.Equals(decimal.TryParse(query, out d) ? d : (decimal?)null))
                     .ToList()
                     .MapToViews();
@@ -64,20 +63,18 @@ namespace TRXpense.App.Web.Controllers
         }
 
         [HttpPost]
-        public ActionResult Save(CountryAllowanceVM countryAllowance)
+        public ActionResult Save(CountryAllowanceVM view)
         {
             ModelState.Remove("Id");
+            if (ModelState.IsValid)
+            {
+                if (view.Id == 0)
+                    _countryAllowanceRepository.AddToDatabase(view.MapToModel());
+                else
+                    _countryAllowanceRepository.UpdateInDatabase(view.MapToModel(), view.Id);
 
-            if (!ModelState.IsValid)
-                return RedirectToAction("Index");
-
-            if (countryAllowance.Id == 0)
-                _countryAllowanceRepository.AddToDatabase(countryAllowance.MapToModel());
-            else
-                _countryAllowanceRepository.UpdateInDatabase(countryAllowance.MapToModel(), countryAllowance.Id);
-
-            _countryAllowanceRepository.Save();
-
+                _countryAllowanceRepository.Save();
+            }
             return RedirectToAction("Index");
         }
 
@@ -85,8 +82,10 @@ namespace TRXpense.App.Web.Controllers
         {
             var allowanceIdDb = _countryAllowanceRepository.FindById(id);
             bool result = false;
+            var travelReportsThatUseThisCountry =
+                _travelReportRepository.GetAllFromDatabaseEnumerable().Where(t => t.CountryAllowanceId == id).ToList().MapToViews();
 
-            if (allowanceIdDb != null)
+            if (travelReportsThatUseThisCountry.Count == 0 && allowanceIdDb != null)
             {
                 _countryAllowanceRepository.DeleteFromDatabase(allowanceIdDb);
                 _countryAllowanceRepository.Save();
